refactor: consolidate type imports from "../types" in event and CSS variable files
